docs, launcher, and final safety audit
- SAFETY_MODEL / RECOVERY / ARCHITECTURE / TESTING / USER_GUIDE docs
- README + run.bat launcher (build + launch desktop)
- Audit: no unwraps on archive input, no silent deletion failures,
  no password leakage into logs, checked arithmetic throughout
- Clippy fixes across crates
- Verified: 66 tests green, release bundles (MSI+NSIS), CLI e2e

+++ crates/platform/src/fs.rs
@@ -37,7 +37,7 @@ pub fn open_for_identity(path: &Path) -> Result<std::fs::File, PlatformError> {
     let result = unsafe {
         CreateFileW(
             PCWSTR(name.as_ptr()),
-            FILE_READ_ATTRIBUTES.0 as u32,
+            FILE_READ_ATTRIBUTES.0,
             FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE,
             None,
             OPEN_EXISTING,
